untested hangman game logic
game.php : main page. displays current hangman stage (1-7), displays word guessed so far, and a letter guessing textbox.

gameLogic.php : function used to validate guesses, function to save guesses in a cookie, function to pick the word to be guessed if there isnt one already saved in a cookie, function update the image on a failed guess, and function to display the correct letters guessed so far.

gameOver.php : page displaying if the player won or lost the current game.